added better handling of API calls in sqlcl script, need to add to rest of code

+++ Invoke_Sqlcl_Session.ps1
@@ -41,16 +41,11 @@ Oracle Database 19c Enterprise Edition Release 19.0.0.0.0 - Production
 Version 19.18.0.1.0
 
 SQL>
+
 ## Invoking script without setting path to sqlcl (sql)
 .\Invoke_Sqlcl_Session.ps1 -BastionId $bastion_ocid -ConnectionId $conn_ocid
 Write-Error: sql not found
-Remove-OpuPortForwardingSessionFull: C:\Users\espenbr\GitHub\oci-powershell-utils\Invoke_Sqlcl_Session.ps1:129
-Line |
- 129 |  … dingSessionFull -BastionSessionDescription $bastionSessionDescription
-     |                                               ~~~~~~~~~~~~~~~~~~~~~~~~~~
-     | Cannot bind argument to parameter 'BastionSessionDescription' because it is null.
 Write-Error: Error: sqlcl not properly installed
-
 
 ## Invoking script with -TestOnly $true
 ❯ .\Invoke_Sqlcl_Session.ps1 -BastionId $bastion_ocid -ConnectionId $conn_ocid -TestOnly $true
@@ -61,6 +56,24 @@ Waiting for creation of bastion session to complete
 Creating SSH tunnel
 DEBUG: Waiting in 30 secs while you check stuff ...
 True
+
+## Invoking script and getting errors and you do not know why (tip: go off VPN)
+❯ .\Invoke_Sqlcl_Session.ps1 -BastionId $bastion_ocid -ConnectionId $adb_conn_ocid
+Getting details from connection
+Creating ephemeral key pair
+Creating Port Forwarding Session to 10.0.1.51:1521
+Waiting for creation of bastion session to complete
+Creating SSH tunnel
+Launching SQLcl
+
+
+SQLcl: Release 22.2 Production on Thu Feb 09 07:40:57 2023
+
+Copyright (c) 1982, 2023, Oracle.  All rights reserved.
+
+  USER          = admin
+  URL           = jdbc:oracle:thin:@tcps://127.0.0.1:9066/hikomo1xnp7z6id_myadb_low.adb.oraclecloud.com?ssl_server_dn_match=off
+  Error Message = IO Error: The Network Adapter could not establish the connection (CONNECTION_ID=sh9DeAMwS9alZ/4KpbQ2DQ==)
 #>
 
 param(
@@ -88,7 +101,7 @@ try {
     }
     ## END: generic section
 
-    ## Make sure mysqlsh is within reach first
+    ## Make sure sqlcl is within reach first
     if ($false -eq (Test-OpuSqlclAvailable)) {
         throw "sqlcl not properly installed"
     }
@@ -99,23 +112,41 @@ try {
     Import-Module OCI.PSModules.Secrets
 
     Out-Host -InputObject "Getting details from connection"
-    ## Grab main handle
-    $connection = Get-OCIDatabasetoolsconnection -DatabaseToolsconnectionId $connectionId
 
-    ## check that this points to an ADB instance
+    ## Grab main handle, ensure it is in correct lifecycle state and that it points to an adb instance
+    try {
+        $connection = Get-OCIDatabasetoolsconnection -DatabaseToolsconnectionId $connectionId -WaitForLifecycleState Active -WaitIntervalSeconds 0 -ErrorAction Stop        
+    }
+    catch {
+        throw "Get-OCIDatabasetoolsconnection: $_"
+    }
     if ("Autonomousdatabase" -ne $connection.RelatedResource.EntityType) {
         throw "Connection does not point to an Autonomous database"
     }
 
-    ## Get adb, service_name and secret based on handle
-    $adb = Get-OCIDatabaseAutonomousDatabase -AutonomousDatabaseId $connection.RelatedResource.Identifier
+    ## Grab adb info based on conn handle, ensure it is in correct lifecycle state
+    try {
+        $adb = Get-OCIDatabaseAutonomousDatabase -AutonomousDatabaseId $connection.RelatedResource.Identifier -WaitForLifecycleState Available -WaitIntervalSeconds 0 -ErrorAction Stop        
+    }
+    catch {
+        throw "Get-OCIDatabaseAutonomousDatabase: $_"
+    }
+
+    ## Get secret (read password) from connection handle
+    try {
+        $secret = Get-OCISecretsSecretBundle -SecretId $connection.UserPassword.SecretId -Stage Current -ErrorAction Stop         
+    }
+    catch {
+        throw "Get-OCISecretsSecretBundle: $_"
+    }
+
+    ## Create connection string
     $fullConnStr = $adb.ConnectionStrings.Low
     $connStr =  $fullConnStr.Substring($fullConnStr.LastIndexOf("/") + 1)
-    $secret = Get-OCISecretsSecretBundle -SecretId $connection.UserPassword.SecretId
- 
+
     ## Assign to local variables for readability
     $userName = $connection.UserName
-    $passwordBase64 = (Get-OCISecretsSecretBundle -SecretId $Secret.SecretId).SecretBundleContent.Content
+    $passwordBase64 = $secret.SecretBundleContent.Content
     $targetHost = $adb.PrivateEndpointip
     $targetPort = 1521
   
@@ -137,8 +168,8 @@ try {
     sql -L ${userName}/${password}@tcps://127.0.0.1:${localPort}/${connStr}?ssl_server_dn_match=off
 }
 catch {
-    ## What else can we do? 
-    Write-Error "Error: $_"
+    ## What else can we do?
+    Write-Error "Error caught: $_"
     return $false
 }
 finally {
@@ -146,8 +177,10 @@ finally {
     ## To Maximize possible clean ups, continue on error 
     $ErrorActionPreference = "Continue"
     
-    ## Request cleanup 
-    Remove-OpuPortForwardingSessionFull -BastionSessionDescription $bastionSessionDescription
+    ## Request cleanup if session object has been created
+    if ($null -ne $bastionSessionDescription) {
+        Remove-OpuPortForwardingSessionFull -BastionSessionDescription $bastionSessionDescription
+    }
 
     ## Finally, unload meodule from memory 
     Set-Location $PSScriptRoot
